Move Singleton, StreamWrapper, NullAsyncResult and Json into Text-Grab.Core
Wave 1a - shared leaves that were blocking Waves 3, 5 and 6 independently.
All four verified dependency-free per invariant 8:

- Singleton.cs and StreamWrapper.cs (WrappingStream) moved as-is; the
  latter constructs NullAsyncResult, which moved in the same commit to
  keep the pair compiling together.
- Json.cs kept its Text_Grab.Helpers namespace unchanged even though it
  physically lived under Utilities/ - not "fixed" to match the folder.

IoUtilities.cs needed a real split. The extension lists and the
Is*File/Is*FileExtension predicates plus ListFilesFoldersInDirectory are
pure string/List<string> logic and moved to Text-Grab.Core as-is.
GetEditorModeForPath and GetOpenContentKindForPath return EtwEditorMode
and OpenContentKind, both still in the app's Enums.cs (that merge is
Wave 2a, not this batch), so they can't move yet. GetContentFromPath and
TryToOpenTextFile also stay app-side - they call OcrUtilities and show
Wpf.Ui/WinForms MessageBoxes. All four impure members moved into a new
Text-Grab/Utilities/FileOpenUtilities.cs facade (the IoUtilities name
had to move with the pure half to avoid a duplicate-type collision once
the app references Core). Updated the four call sites this renamed:
EditTextWindow.xaml.cs, App.xaml.cs, and Tests/FilesIoTests.cs (both
enum-returning test cases).

No files deferred from this batch's list.
