+++ Text-Grab.Core/Models/NullAsyncResult.cs
