+++ Text-Grab.Core/Utilities/Json.cs
