+++ Text-Grab.Core/Utilities/Singleton.cs
