+++ Text-Grab.Core/Utilities/StreamWrapper.cs
